pyproject.toml
[project]
name = "pipeline3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "numpy>=2.3.5",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
]
